refactor datavis.py webapp and config into main.py
add __pycache__ to .gitignore

 add requirements.txt
