+++ .gitignore
@@ -1,3 +1,4 @@
 *.*~
 server.conf
 server.json
+__pycache__
