+++ requirements.txt
@@ -1,0 +1,121 @@
+abstract-rendering==0.5.1
+alabaster==0.7.3
+argcomplete==0.8.9
+astropy==1.0.3
+Babel==1.3
+bcolz==0.9.0
+beautifulsoup4==4.3.2
+binstar==0.11.0
+bitarray==0.8.1
+blaze==0.8.0
+blz==0.6.2
+bokeh==0.9.0
+boto==2.38.0
+Bottleneck==1.0.0
+certifi==14.5.14
+cffi==1.1.0
+clyent==0.3.4
+colorama==0.3.3
+configobj==5.0.6
+cryptography==0.9.1
+Cython==0.22.1
+cytoolz==0.7.3
+datashape==0.4.5
+decorator==4.0.2
+docutils==0.12
+fastcache==1.0.2
+Flask==0.10.1
+geopy==1.11.0
+greenlet==0.4.7
+h5py==2.5.0
+idna==2.0
+ipykernel==4.1.0
+ipython==4.0.0
+ipython-genutils==0.1.0
+ipywidgets==4.0.3
+itsdangerous==0.24
+jdcal==1.0
+jedi==0.8.1
+Jinja2==2.8
+jsonschema==2.5.1
+jupyter==1.0.0
+jupyter-client==4.0.0
+jupyter-console==4.0.3
+jupyter-core==4.0.6
+jupyterhub==0.2.0
+llvmlite==0.5.0
+lxml==3.4.4
+MarkupSafe==0.23
+matplotlib==1.4.3
+mistune==0.7.1
+mock==1.0.1
+multipledispatch==0.4.7
+nbconvert==4.0.0
+nbformat==4.0.1
+nbimporter==0.2
+networkx==1.9.1
+nltk==3.0.3
+nose==1.3.7
+notebook==4.0.5
+numba==0.19.1
+numexpr==2.4.3
+numpy==1.9.2
+odo==0.3.2
+openpyxl==1.8.5
+pandas==0.16.2
+path.py==0.0.0
+patsy==0.3.0
+pep8==1.6.2
+pexpect==3.3
+pickleshare==0.5
+Pillow==2.8.2
+ply==3.6
+psutil==2.2.1
+ptyprocess==0.4
+py==1.4.27
+pyasn1==0.1.7
+pycosat==0.6.1
+pycparser==2.14
+pycrypto==2.6.1
+pycurl==7.19.5.1
+pyflakes==0.9.2
+Pygments==2.0.2
+pyOpenSSL==0.15.1
+pyparsing==2.0.3
+pytest==2.7.1
+python-dateutil==2.4.2
+pytz==2015.4
+PyYAML==3.11
+pyzmq==14.7.0
+qtconsole==4.0.1
+redis==2.10.3
+requests==2.8.0
+rope-py3k==0.9.4.post1
+runipy==0.1.3
+scikit-image==0.11.3
+scikit-learn==0.16.1
+scipy==0.16.0
+simplegeneric==0.8.1
+simplekml==1.2.8
+simplepam==0.1.5
+six==1.9.0
+snowballstemmer==1.2.0
+sockjs-tornado==1.0.1
+Sphinx==1.3.1
+sphinx-rtd-theme==0.1.7
+spyder==2.3.5.2
+SQLAlchemy==1.0.8
+statsmodels==0.6.1
+sympy==0.7.6
+tables==3.2.0
+terminado==0.5
+Theano==0.7.0
+toolz==0.7.2
+tornado==4.2.1
+traitlets==4.0.0
+ujson==1.33
+Werkzeug==0.10.4
+wheel==0.24.0
+xlrd==0.9.3
+XlsxWriter==0.7.3
+xlwt==1.0.0
